Relocate the Counter functions into separate modules aand update the meal counter test
