+++ src/Modules/commentsCounter.js
@@ -1,8 +1,3 @@
-const mealCounter = (arr) => {
-  const navDishes = document.querySelector('#nav-dishes');
-  navDishes.textContent += ` (${arr.length})`;
-};
-
 const commentsCounter = (arr, header) => {
   if (Array.isArray(arr)) {
     header.textContent = `Comments (${arr.length})`;
@@ -10,4 +5,4 @@ const commentsCounter = (arr, header) => {
     header.textContent = 'Comments (0)';
   }
 };
-export { mealCounter, commentsCounter };
+export default commentsCounter;
